fix: Add foot_util.dart
